refactor: Separated formik hooks to a separate file
