+++ src/hooks/useRegisterUser.tsx
@@ -1,0 +1,18 @@
+import { useMutation } from "react-query";
+import { useRouter } from "next/router";
+
+import { loginUserFn } from "../utils/authApi";
+import { LoginInput } from "../types/LoginInput";
+
+export default function useRegisterUser() {
+  const router = useRouter();
+  const mutation = useMutation(
+    (userData: LoginInput) => loginUserFn(userData),
+    {
+      onSuccess: () => {
+        router.push("/");
+      },
+    }
+  );
+  return mutation;
+}
